refactored CLI
