+++ packages/hpe-design-tokens/src/scripts/figma-api-cli/index.ts
@@ -1,7 +1,7 @@
-import CLI from './figma-api-cli.js';
+import cli from './figma-api-cli.js';
 import { brightRed } from '../../utils.js';
 
-CLI().catch(error => {
+cli().catch(error => {
   const message = error instanceof Error ? error.message : String(error);
   console.error(brightRed(`Fatal error: ${message}`));
   process.exit(1);
